feat(globalhooks): add the ability to access a copy of the features metadata
A global hook (BeforeAll and AfterAll) will run before or after every
single feature.
However they could only be used to add information to the context (which
is empty when the feature starts).
No information about the feature could be used in the hooks. The global
hooks are now capable of passing a copy of the metadata (name of the
feature and tags, at the moment) to the hook function, allowing for more
targeted behaviors.