chore(dependencies): upgrade to yarn3, update dependencies to latest versions, and adapt peerdeps
yarn3 is now the default package manager for this package. Cucumber dependencies have been updated
to the latest major version (8). Peer dependencies have been adapted to prevent the installation of
testcafe 1.19.0 (seems to work but issues with some of its dependencies, waiting for a more stable
release) and to force the update of cucumber. Also added cucumber peer dependencies to the package
to avoid warnings (they don't seem to be needed)

fix #77